[BUGFIX] Only override default doctrine types when needed
This package changes the default doctrine types with a custom
map to archieve the goal when generating model files. However,
this interferes with other packages, needing to deal with real
schemaninformation. They get wrong database schema information
with the changed type registration.

Therefore, this change moves the registration of default type
map in the `\Krlove\EloquentModelGenerator\TypeRegistry` from
the constructor to a dedicated method guarded with a flag to
avoid re-registering it. This new method to register default
types is called only for commands needing this and not when it
is invoked in general.

This is a slightly other approach then #106 to solve #103. The
main difference is, that constructor injection is kept as is.

Resolves #103
Related #106